# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AirGo3D Backend API",
    "description": "API documentation for AirGo3D backend service. Provides endpoints for S3 presigned URL generation and file management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "S3",
      "description": "AWS S3 presigned URL generation endpoints"
    },
    {
      "name": "Panorama",
      "description": "PanoramaImage management endpoints"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Returns a simple health check message to verify the API is running",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "text/plain": {
                "example": "Hello"
              }
            }
          }
        }
      }
    },
    "/api/s3/presigned-url": {
      "post": {
        "tags": ["S3"],
        "summary": "Generate presigned URL for S3 upload",
        "description": "Generates a presigned URL that allows clients to upload files directly to S3 without exposing AWS credentials. The S3 object key is automatically generated as a UUID v4. The URL expires after a specified time (default: 5 minutes).",
        "operationId": "generatePresignedUrl",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contentType": {
                    "type": "string",
                    "description": "MIME type of the file being uploaded",
                    "example": "image/jpeg"
                  },
                  "expiresIn": {
                    "type": "integer",
                    "description": "Expiration time for the presigned URL in seconds (default: 300 seconds / 5 minutes)",
                    "minimum": 1,
                    "maximum": 604800,
                    "default": 300,
                    "example": 300
                  },
                  "prefix": {
                    "type": "string",
                    "description": "Optional prefix/path for the S3 key. The UUID will be appended to this prefix (e.g., 'uploads/images/' will result in 'uploads/images/{uuid}')",
                    "example": "uploads/images"
                  }
                }
              },
              "examples": {
                "imageUpload": {
                  "summary": "Image upload example with prefix",
                  "value": {
                    "contentType": "image/jpeg",
                    "prefix": "uploads/images",
                    "expiresIn": 300
                  }
                },
                "documentUpload": {
                  "summary": "Document upload example",
                  "value": {
                    "contentType": "application/pdf",
                    "prefix": "uploads/documents",
                    "expiresIn": 600
                  }
                },
                "minimalRequest": {
                  "summary": "Minimal request (all fields optional)",
                  "value": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "description": "The presigned URL for uploading the file to S3",
                      "example": "https://s3.amazonaws.com/bucket-name/uploads/images/photo-123.jpg?X-Amz-Algorithm=..."
                    },
                    "key": {
                      "type": "string",
                      "description": "The auto-generated S3 object key (UUID v4). If a prefix was provided, it will be in the format '{prefix}/{uuid}'",
                      "example": "uploads/images/550e8400-e29b-41d4-a716-446655440000"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "Expiration time in seconds",
                      "example": 300
                    }
                  },
                  "required": ["url", "key", "expiresIn"]
                },
                "example": {
                  "url": "https://s3.amazonaws.com/bucket-name/uploads/images/550e8400-e29b-41d4-a716-446655440000?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=...",
                  "key": "uploads/images/550e8400-e29b-41d4-a716-446655440000",
                  "expiresIn": 300
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Failed to generate presigned URL",
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/api/panorama": {
      "post": {
        "tags": ["Panorama"],
        "summary": "Create a new PanoramaImage",
        "description": "Creates a new PanoramaImage record in the database after a successful S3 upload. The filename is automatically extracted from the S3 key.",
        "operationId": "createPanoramaImage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePanoramaImageRequest"
              },
              "examples": {
                "imageWithMetadata": {
                  "summary": "Image with full metadata",
                  "value": {
                    "key": "images/photo-123-550e8400-e29b-41d4-a716-446655440000",
                    "name": "Beautiful Panorama",
                    "originalFilename": "panorama.jpg",
                    "fileSize": 2048576,
                    "mimeType": "image/jpeg",
                    "width": 3840,
                    "height": 2160,
                    "metadata": {
                      "camera": "Canon EOS R5",
                      "location": "Mountain View",
                      "tags": ["nature", "panorama", "landscape"],
                      "description": "A beautiful panoramic view"
                    }
                  }
                },
                "minimalImage": {
                  "summary": "Minimal required fields",
                  "value": {
                    "key": "images/photo-123-550e8400-e29b-41d4-a716-446655440000",
                    "name": "My Panorama",
                    "originalFilename": "panorama.jpg",
                    "fileSize": 1024000,
                    "mimeType": "image/jpeg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PanoramaImage created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaImageResponse"
                },
                "example": {
                  "_id": "507f1f77bcf86cd799439011",
                  "name": "Beautiful Panorama",
                  "filename": "photo-123-550e8400-e29b-41d4-a716-446655440000",
                  "originalFilename": "panorama.jpg",
                  "filePath": "images/photo-123-550e8400-e29b-41d4-a716-446655440000",
                  "fileSize": 2048576,
                  "mimeType": "image/jpeg",
                  "width": 3840,
                  "height": 2160,
                  "isBookmarked": false,
                  "uploadedAt": "2024-01-15T10:30:00.000Z",
                  "updatedAt": "2024-01-15T10:30:00.000Z",
                  "metadata": {
                    "camera": "Canon EOS R5",
                    "location": "Mountain View",
                    "tags": ["nature", "panorama", "landscape"],
                    "description": "A beautiful panoramic view"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "missingField": {
                    "value": {
                      "error": "Missing or invalid 'key' field. 'key' must be a non-empty string."
                    }
                  },
                  "invalidMimeType": {
                    "value": {
                      "error": "Invalid 'mimeType'. Must be one of: image/jpeg, image/png, image/jpg, image/webp"
                    }
                  },
                  "duplicateFilename": {
                    "value": {
                      "error": "A PanoramaImage with this filename already exists.",
                      "message": "E11000 duplicate key error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Failed to create PanoramaImage",
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/api/panorama/{id}/bookmark": {
      "patch": {
        "tags": ["Panorama"],
        "summary": "Update bookmark status of a PanoramaImage",
        "description": "Updates the bookmark status (bookmark or unbookmark) of a PanoramaImage by its ID.",
        "operationId": "updatePanoramaImageBookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB document ID of the PanoramaImage",
            "schema": {
              "type": "string"
            },
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isBookmarked"],
                "properties": {
                  "isBookmarked": {
                    "type": "boolean",
                    "description": "Bookmark status (true to bookmark, false to unbookmark)",
                    "example": true
                  }
                }
              },
              "examples": {
                "bookmark": {
                  "summary": "Bookmark the image",
                  "value": {
                    "isBookmarked": true
                  }
                },
                "unbookmark": {
                  "summary": "Unbookmark the image",
                  "value": {
                    "isBookmarked": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bookmark status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaImageResponse"
                },
                "example": {
                  "_id": "507f1f77bcf86cd799439011",
                  "name": "Beautiful Panorama",
                  "filename": "photo-123-550e8400-e29b-41d4-a716-446655440000",
                  "originalFilename": "panorama.jpg",
                  "filePath": "images/photo-123-550e8400-e29b-41d4-a716-446655440000",
                  "fileSize": 2048576,
                  "mimeType": "image/jpeg",
                  "width": 3840,
                  "height": 2160,
                  "isBookmarked": true,
                  "uploadedAt": "2024-01-15T10:30:00.000Z",
                  "updatedAt": "2024-01-15T10:35:00.000Z",
                  "metadata": {
                    "camera": "Canon EOS R5",
                    "location": "Mountain View",
                    "tags": ["nature", "panorama", "landscape"],
                    "description": "A beautiful panoramic view"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "missingField": {
                    "value": {
                      "error": "Missing or invalid 'isBookmarked' field. 'isBookmarked' must be a boolean."
                    }
                  },
                  "invalidId": {
                    "value": {
                      "error": "Invalid PanoramaImage ID format."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "PanoramaImage not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "PanoramaImage not found."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Failed to update bookmark status",
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message (optional)"
          }
        },
        "required": ["error"]
      },
      "CreatePanoramaImageRequest": {
        "type": "object",
        "required": ["key", "name", "originalFilename", "fileSize", "mimeType"],
        "properties": {
          "key": {
            "type": "string",
            "description": "S3 object key/path from the upload",
            "example": "images/photo-123-550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "Display name for the image",
            "example": "Beautiful Panorama"
          },
          "originalFilename": {
            "type": "string",
            "description": "Original filename from client",
            "example": "panorama.jpg"
          },
          "fileSize": {
            "type": "integer",
            "description": "File size in bytes",
            "minimum": 0,
            "example": 2048576
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the image",
            "enum": ["image/jpeg", "image/png", "image/jpg", "image/webp"],
            "example": "image/jpeg"
          },
          "width": {
            "type": "integer",
            "description": "Image width in pixels",
            "minimum": 0,
            "example": 3840
          },
          "height": {
            "type": "integer",
            "description": "Image height in pixels",
            "minimum": 0,
            "example": 2160
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata",
            "properties": {
              "camera": {
                "type": "string",
                "example": "Canon EOS R5"
              },
              "location": {
                "type": "string",
                "example": "Mountain View"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["nature", "panorama", "landscape"]
              },
              "description": {
                "type": "string",
                "example": "A beautiful panoramic view"
              }
            }
          }
        }
      },
      "PanoramaImageResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB document ID",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "Beautiful Panorama"
          },
          "filename": {
            "type": "string",
            "description": "Extracted filename from S3 key",
            "example": "photo-123-550e8400-e29b-41d4-a716-446655440000"
          },
          "originalFilename": {
            "type": "string",
            "example": "panorama.jpg"
          },
          "filePath": {
            "type": "string",
            "description": "S3 object key/path",
            "example": "images/photo-123-550e8400-e29b-41d4-a716-446655440000"
          },
          "fileSize": {
            "type": "integer",
            "example": 2048576
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "width": {
            "type": "integer",
            "example": 3840
          },
          "height": {
            "type": "integer",
            "example": 2160
          },
          "isBookmarked": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "camera": {
                "type": "string"
              },
              "location": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "description": {
                "type": "string"
              }
            }
          }
        },
        "required": ["_id", "name", "filename", "originalFilename", "filePath", "fileSize", "mimeType", "isBookmarked", "uploadedAt", "updatedAt"]
      }
    }
  }
}